requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pymysql
mysql-connector-python
python-dotenv